move some to fabric-manager component
