+++ components/accelerator/nvidia/fabric-manager/testdata/lspci.nn.a10
