+++ components/accelerator/nvidia/fabric-manager/testdata/nvidia-smi.nvlink.status.a100
